refactor: ajustando Playground para diretório correto
